Bugfix: Viewing Network when Signed In
Viewing a network, e.g. at `network/?id=1`, was failing because in the
networks controllers, we were getting the user ID with
`current_user.get_id()`. However, the `get_id()` method returns a
dictionary of the user object (as described in docstrings). Instead, we
want to use `current_user.id`.